Cap number of effective appearances to 10
Due to efficiency reasons (i.e. there are 2^n subsets of n appearances) lowered the effective appearance cap from 64 to 10. This was done by:
- Only showing the first 10 forms for pku slots
- Limiting the DexUtil.GetSearchableAppearances method to 10 appearances.

Note that pku files can still have more appearances in the appearance array, but only the first 10 will be recognized by pkuManager.

Not an ideal solution, but fine for now...