Final commit
With this empty commit, this iteration of pkuManager will be deprecated. It will be succeeded by pkuManager v2 which is being developed in flutter.

The use of WinForms as a UI was... regrettable, and the need for a replacement was apparent. Further, in my quest to make adding new formats painless, I created a tangled web of needless abstraction...

Since both the UI and model needed to be completely rewritten. there was no logner anything binding the project to be written in C#. Hopefully, however, the lessons learned from the many rewrites of this pkuManager will not be lost.